NMLab_final-main/requirements.txt
asgiref==3.4.1
boto3==1.24.3
botocore==1.27.3
certifi==2022.5.18.1
click==8.1.3
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
Flask==2.1.2
future==0.18.2
iso8601==1.0.2 
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
paho-mqtt==1.5.1
protobuf==3.18.0
pynmea2==1.18.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.1
pyyaml==6.0
s3transfer==0.6.0
schedule==1.1.0
serial==0.0.97
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.9
Werkzeug==2.1.2
